Started a Resource Manager, so I can load in resources used to test the OpenGL Draw Functions

+++ Project Diode/Project Diode/shaders.h
@@ -24,19 +24,19 @@ enum class ShaderType : unsigned long long {
   Count
 };
 
-enum class VertexShaders : unsigned {
+enum class VertexShaders : unsigned long long {
   Default = 0,
 };
 
-enum class TessellationShaders : unsigned {
+enum class TessellationShaders : unsigned long long {
   Default = 2,
 };
 
-enum class GeometryShaders : unsigned {
+enum class GeometryShaders : unsigned long long {
   Default = 4,
 };
 
-enum class FragmentShaders : unsigned {
+enum class FragmentShaders : unsigned long long {
   Default = 8,
 };
